modifying user related files to use authMiddleware

+++ src/controllers/UserController.ts
@@ -1,45 +1,30 @@
 import type { Request, Response } from "express"
 import bcrypt from "bcryptjs"
 import User from "../models/user"
-import { createUserSchema, updateUserSchema, userIdSchema } from "../schemas/userSchema"
+import { updateUserSchema, userIdSchema } from "../schemas/userSchema"
 import { z } from "zod"
+import type { AuthRequest } from "../middlewares/authmiddleware"
 
 export const getUsers = async (req: Request, res: Response) => {
     try {
-        const users = await User.find().select("-password")
+        const users = await User.find().select("name email -_id")
         res.status(200).json(users)
     } catch (error: any) {
         res.status(500).json({ message: error.message || "Something went wrong" })
     }
 }
 
-export const createUser = async (req: Request, res: Response) => {
-    try {
-        const validatedData = createUserSchema.parse(req.body)
-        const hashedPassword = await bcrypt.hash(validatedData.password, 10)
-        const user = new User({ ...validatedData, password: hashedPassword })
-        await user.save()
-        res.status(201).json({
-            message: "User created successfully",
-            user: { ...user.toObject(), password: undefined }
-        })
-    } catch (error: any) {
-        if (error instanceof z.ZodError) {
-            return res.status(400).json({
-                message: error.issues.map(issue => issue.message)})
-        }
-        res.status(500).json({message: error.message || "Something went wrong"})
-    }
-}
-
-export const getUserById = async (req: Request, res: Response) => {
+export const getUserById = async (req: AuthRequest, res: Response) => {
     try {
         const id = userIdSchema.parse(req.params.id)
-        const user = await User.findById(id)
+        if (req.user._id.toString() !== id) {
+            return res.status(403).json({message: "Access denied"})
+        }
+        const user = await User.findById(id).select("-password")
         if (!user) {
             return res.status(404).json({message: "User not found"})
         }
-        res.status(200).json({ ...user.toObject(), password: undefined })
+        res.status(200).json(user)
     } catch (error: any){
         if (error instanceof z.ZodError) {
             return res.status(400).json({
@@ -49,9 +34,12 @@ export const getUserById = async (req: Request, res: Response) => {
     }
 }
 
-export const updateUser = async (req: Request, res: Response) => {
+export const updateUser = async (req: AuthRequest, res: Response) => {
     try {
         const id = userIdSchema.parse(req.params.id)
+        if (req.user._id.toString() !== id) {
+            return res.status(403).json({message: "Access denied"})
+        }
         const user = await User.findById(id)
         if (!user) {
             return res.status(404).json({message: "User not found"})
@@ -60,7 +48,7 @@ export const updateUser = async (req: Request, res: Response) => {
         if (validatedData.name) user.name = validatedData.name
         if (validatedData.email) user.email = validatedData.email
         if (validatedData.password) {
-            validatedData.password  = await bcrypt.hash(validatedData.password, 10)
+            user.password  = await bcrypt.hash(validatedData.password, 10)
         }
         await user.save()
         res.status(200).json({ 
@@ -77,9 +65,12 @@ export const updateUser = async (req: Request, res: Response) => {
     }
 }
 
-export const deleteUser = async (req: Request, res: Response) => {
+export const deleteUser = async (req: AuthRequest, res: Response) => {
     try {
         const id = userIdSchema.parse(req.params.id)
+        if (req.user._id.toString() !== id) {
+            return res.status(403).json({message: "Access denied"})
+        }
         const user = await User.findByIdAndDelete(id)
         if (!user) {
             return res.status(404).json({message: "User not found"})
